TOOLS: Add configuration to split into 2 separate packages
Webpack will now output 2 separate packages

1. Shredder API
2. Shredder (contains both UI and API)

If Shredder package is loaded the UI will bootstrap and add itself to the page while containing also the api in its bundle.

If the script is included via script tags in the page there will be a global called Shredder.api available for use.