Treat requests params as strings.
An HTTP request is combined of string fields.
We should not make any assumption about the type of content of an HTTP request not in the response body or the request body.

This is done because of the following reasons:
1. We could mess up with the type assumption and break a request (which was this use case).
2. We should treat a request as its original value, if we wish to change that value then we need to respect its original structure.
This means that if we want to we can handle different content types in the UI in order to display the parameters in a cleaner way for specific supported types and avoid trying to support all of them because that is not possible without incredible amount of work.
3. Treating this as a string will help us going forward dynamic templates for responses as we will only need to compile the end result back to a string regardless of the content type.