fix(export): Remove an incorrect check of the content type to attempt parsing request parameters
Before this fix, only mocks that returned a content type of application/json would parse the request

parameters and request body due to a wrong assumption. if a request was sent using json request

parameters (json body) but the server would return a different response type, such as text/html the

check will fail and return an incorrect export file.