Release v1.0.3 (#148)
* Update xhook dependency with modified support for web workers

* Add an abstraction between the interceptor and mock data

* Add ability to mock a fetch request inside a web worker

* bump version to v1.0.3-beta.1

* Change interval to a timeout and rename data property to payload

* bump version to v1.0.3-beta.2

* Fix bdsm not updating content with correct message type

* Register an event listener on the worker and if it belongs to bdsm stop the event from propagating

* Namespace postmessage events and stop event propagation on bdsm messages on the client as well

* restore normal requests

* Add bdsm/worker entry point

* Add bdsm/api entry point

* feat(Export): Add new export dialog with ability to prettify exports (#143)

* feat(API): Add ability to prettify exports in the API

Add a flag to the export function that allows you to get a prettified response. New method signature

is export(prettify:bool). If prettyify is supplied the output will be a json string with 2 spaces

indentation

* Implement new export dialog

* Tests: Add Interceptor E2E tests

Add e2e tests to all use cases of requests being captured and modified via the interceptor

* fix(export): Remove an incorrect check of the content type to attempt parsing request parameters (#146)

Before this fix, only mocks that returned a content type of application/json would parse the request

parameters and request body due to a wrong assumption. if a request was sent using json request

parameters (json body) but the server would return a different response type, such as text/html the

check will fail and return an incorrect export file.

* chore(scripts): Streamline the npm run test:integration command (#147)

the test:integration command now builds the project, opens an http server, runs the tests and kills

the http server process in the end instead of having to do these steps manually

* chore(Release): Release version 1.0.3